requirements.txt
Flask==2.1.2
Keras_Preprocessing==1.1.2
numpy==1.22.4
Pillow==9.1.1
tensorflow==2.9.1
gunicorn==20.1.0